pyproject.toml
[project]
name = "track_remote_sensing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "FergusCurrie", email = "fergus.currie@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "albumentations>=2.0.8",
    "folium>=0.20.0",
    "hvplot>=0.12.1",
    "ipykernel>=6.30.1",
    "ipython>=8.37.0",
    "ipywidgets>=8.1.7",
    "label-studio-sdk>=2.0.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "mercantile>=1.2.1",
    "onnx>=1.19.0",
    "opencv-python>=4.12.0.88",
    "poethepoet>=0.37.0",
    "polars>=1.33.0",
    "pre-commit>=4.3.0",
    "pyproj>=3.7.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.12",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "segment-anything-py>=1.0.1",
    "segmentation-models-pytorch>=0.5.0",
    "shapely>=2.1.1",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
]

[project.scripts]
track_remote_sensing = "track_remote_sensing:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.poe.tasks]
test         = "echo hi"                         # a simple command task
expand-images = "python3 src/track_remote_sensing/main.py --mode expand_imagery"
push-images = "python3 src/track_remote_sensing/main.py --mode push_imagery"
pull-labels = "python3 src/track_remote_sensing/main.py --mode pull_labels"


[tool.ruff.format]
indent-style = "tab"
